[ex02] starting to rewrite code after discovering some issues with mohamad. Had a mind bogging issue with the new dummy_val functions but i finally fixed it

+++ ex02/PmergeMe.hpp
@@ -6,7 +6,7 @@
 /*   By: mhuszar <mhuszar@student.42vienna.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/30 20:29:10 by mhuszar           #+#    #+#             */
-/*   Updated: 2024/09/10 16:38:57 by mhuszar          ###   ########.fr       */
+/*   Updated: 2024/09/10 17:38:34 by mhuszar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 #include <iostream>
 #include <deque>
 
-class PmergeMe : public IMerge< std::deque<int> >
+class PmergeMe : public IMerge< std::deque<long int> >
 {
     public:
